feat: zip with folders
zip now has two functions, root and not root.
root will look at the archive and attempty to find the
true root of the archive (for those archives that are
zips of folders) so when you are extracting it you are
always getting the contents of the archive and not just
a folder containing the contents

also creates some tests with test archives to test these,
with cleanup to remove the extracted files after passing,
if the test fails then you will have to manually clean the
testing directory, be careful not to commit any of the test
files (though they are empty, so shouldn't be too bad if you
do).