fix: solves failed tests
solves failing tests for all three formats, not sure how i commited it
and released it before with these failings because they were fairly big.

- fixed issues with `tar` where it couldn't extract a file because that
file required a folder and no folder was created, so now checks if there
are parent folders and makes sure they exist, and if they don't it
creates them.
- fixes issue with `zip` where it forgets to add the file name to a file
that is in the root of the archive so it fails when trying to extact
because the name of the file equals the root folder where it is trying
to extract.
- reworks the tests slightly so that logging is enabled while performing
a test, helps find out the root cause for some of these failures because
most of the user facing errors are expecting the insides to work.
- reworks the test files and makes them all unique. tests would
occasionally fail because (i'm guessing) the tests would run
concurrently and one test would remove the file for the other test,
causing the test to fail because the file didnt exist that time.
assuming this is the case because running tests multiple times would
result in unrepeatable results, uniquely naming each file has fix this
and now they pass 100% of the time.