feat: adds create archive, zip only for now
- adds create module with a function to create a zip file
- creates tests to confirm the zip is created correctly
- adds function in read to read the contents of a file, used to confirm
that the create zip method works correctly
- changes some signature functions, read below for more information.

changes most function signatures from `<P : AsRef<Path>>` to
`<P1 : AsRef<Path>, P2 : AsRef<Path>>`. This now allows you to
enter the path information as two different types. If you use the
`P` method than both the `src` and the `des` need to be the same
type. This doesn't always make sense, for example if you wanted to
extract everything to a temp directory, you wouldn't be able to do
this

    archive::extract::root_to(archive_as_a_path, "temp")?;

instead you would need to do this

    archive::extract::root_to(archive_as_a_path_buf,PathBuf::from("temp"))?;