refactor: reorganizes the archive create for zip function
makes a new module and preps for creation of other formats.

has an issue with tests failing randomly, need to figure out why
and then figure out how to make it more repeatable. assuming its
the same issue as before, something to do with creating and deleting
test stuff. i think there is a crate that helps with this, investigate.